Add a clean wordpress migration
The migration contains everything needed for a newly installed
wordpress

The mysql service used in the docker-compose.yml file will run
all migrations inside the infrastructure/migrations directory
when the service starts and doesn't find any database.

To trigger a full migration again you can remove everything inside
the infrastructure/data/mysql directory, but beware of data loss.
You should make a backup before if you want your data.

When the database already exists, if you add new migrations to the
directory, you would need to run them manually with mysqldump or
some alternative.

Adding a library to handle database migrations is being thought
but this is good for now.

The first time the mysql service starts it will do the
following steps:
  - Set the root password
  - Create a new database
  - Create a new user with password and assign privileges on the
    newly created database
  - Run all sql files found in infrastructure/migrations/*.sql

Every setting related to the database, users and passwords can be
configured in the .env file (make a copy of the .env.example if
you don't have a .env). Just do it before running the mysql service